+++ sessions-app/src/main/java/com/mironov/sessions_app/DTO/LobbyMemberDTO.java
@@ -1,17 +1,17 @@
 package com.mironov.sessions_app.DTO;
 
 public class LobbyMemberDTO {
-    private Integer id;
-    private LobbyDTO lobby; // Включаем DTO для лобби
-    private UserDTO user; // Включаем DTO для пользователя
+    private Long id;
+    private LobbyDTO lobby;
+    private UserDTO user;
     private String joinedAt;
 
     // Геттеры и сеттеры
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
