+++ sessions-app/src/main/java/com/mironov/sessions_app/entity/LobbyEntity.java
@@ -7,7 +7,7 @@ public class LobbyEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -17,12 +17,11 @@ public class LobbyEntity {
 
     private String createdAt;
 
-    // Геттеры и сеттеры
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
