+++ sessions-app/src/main/java/com/mironov/sessions_app/entity/LobbyMemberEntity.java
@@ -7,7 +7,7 @@ public class LobbyMemberEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "lobby_id")
@@ -19,12 +19,11 @@ public class LobbyMemberEntity {
 
     private String joinedAt;
 
-    // Геттеры и сеттеры
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
