+++ sessions-app/src/main/java/com/mironov/sessions_app/entity/MessageEntity.java
@@ -8,7 +8,7 @@ public class MessageEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "lobby_id")
@@ -24,12 +24,11 @@ public class MessageEntity {
     @JoinColumn(name = "user_id")
     private UserEntity user;
 
-    // Геттеры и сеттеры
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
