+++ sessions-app/src/main/java/com/mironov/sessions_app/entity/UserEntity.java
@@ -7,7 +7,7 @@ public class UserEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(unique = true)
     private String email;
@@ -15,12 +15,11 @@ public class UserEntity {
     private String username;
     private String createdAt;
 
-    // Геттеры и сеттеры
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
