+++ sessions-app/src/main/java/com/mironov/sessions_app/entity/UserGameEntity.java
@@ -8,7 +8,7 @@ public class UserGameEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
@@ -20,12 +20,11 @@ public class UserGameEntity {
 
     private String rankValue;
 
-    // Геттеры и сеттеры
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
